# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ECHO Mind Path A API",
    "description": "心理健康记录、筛查提示、审核练习、人工接管与审计。非诊断、非紧急服务替代。",
    "version": "0.6.0"
  },
  "paths": {
    "/v1/tenants": {
      "post": {
        "summary": "Create Tenant",
        "operationId": "create_tenant_v1_tenants_post",
        "parameters": [
          {
            "name": "x-bootstrap-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Bootstrap-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users": {
      "post": {
        "summary": "Create User",
        "operationId": "create_user_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/onboarding/l0": {
      "post": {
        "summary": "Create L0",
        "operationId": "create_l0_v1_onboarding_l0_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/L0ScreeningCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/onboarding/emergency-contact": {
      "post": {
        "summary": "Create Emergency Contact",
        "operationId": "create_emergency_contact_v1_onboarding_emergency_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyContactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/onboarding/verify-code": {
      "post": {
        "summary": "Verify Onboarding Code",
        "description": "激活码交换（v0.6.1）：预认证薄路由，供端侧激活使用。\n\n主路径：机构激活码（ActivationCode）——哈希存储 / TTL / 一次性消费 /\n防爆破 rate limit / IP+device 审计；revoked/expired/replay/restricted/超限 → 403。\nlegacy 回退（v0.8 退役）：未命中时回退查 User.external_ref == code。\n响应不含 tenant_id / role / external_ref 等内部字段。",
        "operationId": "verify_onboarding_code_v1_onboarding_verify_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingVerifyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/profile/{user_id}": {
      "get": {
        "summary": "Get Profile",
        "description": "只读画像缓存：绝不 rebuild / version+1 / commit（无任何写副作用）。",
        "operationId": "get_profile_v1_profile__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profile/{user_id}/rebuild": {
      "post": {
        "summary": "Rebuild User Profile",
        "description": "显式重建画像：traits + version+1；仅写路径调用。",
        "operationId": "rebuild_user_profile_v1_profile__user_id__rebuild_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sandbox/runs": {
      "post": {
        "summary": "Schedule Sandbox",
        "description": "触发一次自进化沙箱运行（幂等：同 tenant+user+date 返回同一记录）。\n\n并发配额：调度期 running 计数快速失败 + 执行期原子租户槽（双保险）。",
        "operationId": "schedule_sandbox_v1_sandbox_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/sandbox/runs/{run_id}": {
      "get": {
        "summary": "Get Sandbox Run",
        "description": "查询单个沙箱运行记录。受 ensure_user 校验目标用户归属。",
        "operationId": "get_sandbox_run_v1_sandbox_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmergencyContactCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "minLength": 5,
            "title": "Phone"
          },
          "relationship": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Relationship"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "phone",
          "relationship"
        ],
        "title": "EmergencyContactCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "L0ScreeningCreate": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "title": "Event Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "current_danger": {
            "type": "boolean",
            "title": "Current Danger",
            "default": false
          },
          "prior_attempt_or_admission": {
            "type": "boolean",
            "title": "Prior Attempt Or Admission",
            "default": false
          },
          "psychosis_or_mania": {
            "type": "boolean",
            "title": "Psychosis Or Mania",
            "default": false
          },
          "substance_impairment": {
            "type": "boolean",
            "title": "Substance Impairment",
            "default": false
          },
          "has_professional_support": {
            "type": "boolean",
            "title": "Has Professional Support",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "user_id"
        ],
        "title": "L0ScreeningCreate"
      },
      "OnboardingVerifyIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "OnboardingVerifyIn",
        "description": "激活码交换入参（v0.6 final）。\n\n仅接受机构分发的激活码/邀请码；不接收 device_id 等内部标识之外的字段。\ncode 最小 8 位（如 `XXXX-XXXX-XXXX`），最长 80。"
      },
      "OnboardingVerifyOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "consent_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Consent Versions"
          },
          "l0_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "L0 Decision"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "access_token"
        ],
        "title": "OnboardingVerifyOut",
        "description": "激活码交换输出契约（v0.6 final）。\n\n仅返回端侧所需字段；不暴露 tenant_id / role / external_ref / bootstrap 等\n内部字段（JWT 载荷内部字段对用户透明）。"
      },
...
      "SandboxRunCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "run_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Date"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "SandboxRunCreate",
        "description": "触发沙箱运行的入参。run_date 缺省为当天（UTC）。"
      },
...
      "TenantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TenantCreate"
      },
...
      "UserCreate": {
        "properties": {
          "external_ref": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "External Ref"
          },
          "age_band": {
            "type": "string",
            "const": "18_plus",
            "title": "Age Band",
            "default": "18_plus"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Shanghai"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          }
        },
        "type": "object",
        "required": [
          "external_ref"
        ],
        "title": "UserCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}